MyPageController edit -> SharedGameController
