+++ src/main/java/com/scriptopia/demo/controller/MyPageController.java
@@ -21,16 +21,6 @@ public class MyPageController {
 
 
     /*
-    게임 공유 -> 게임 공유하기
-     */
-    @PostMapping("/my-page/share/{uuid}")
-    public ResponseEntity<?> share(Authentication authentication, @PathVariable UUID uuid) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        return sharedGameService.saveSharedGame(userId, uuid);
-    }
-
-    /*
     유저 -> 사용자 게임 기록 조회
      */
     @GetMapping("/my-page/history")
@@ -42,28 +32,6 @@ public class MyPageController {
         return historyService.fetchMyHistory(userId, lastId, size);
     }
 
-    /*
-    게임 공유 -> 공유한 게임 조회(내가 공유한 게임 조회)
-     */
-    @GetMapping("/my-page/games/shared")
-    public ResponseEntity<?> getMySharedGames(Authentication authentication) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        return sharedGameService.getMySharedGames(userId);
-    }
-
-
-    /*
-    게임 공유 -> 공유한 게임 삭제
-     */
-    @DeleteMapping("/my-page/share/{uuid}")
-    public ResponseEntity<?> delete(Authentication authentication, @PathVariable UUID uuid) {
-        Long userId = Long.valueOf(authentication.getName());
-
-        sharedGameService.deleteSharedGame(userId, uuid);
-
-        return ResponseEntity.ok("게임이 삭제되었습니다.");
-    }
 
     /*
     게임 -> 기존 게임 조회
@@ -73,5 +41,15 @@ public class MyPageController {
         Long userId = Long.valueOf(authentication.getName());
 
         return gameSessionService.getGameSession(userId);
+    }
+
+    /*
+    게임 -> 기존 게임 삭제
+     */
+    @DeleteMapping("/my-page/game/{gameId}")
+    public ResponseEntity<?> deleteGameSession(Authentication authentication, @PathVariable String gameId) {
+        Long userId = Long.valueOf(authentication.getName());
+
+        return gameSessionService.deleteGameSession(userId, gameId);
     }
 }
